Refacture codebase to use modules
